+++ src/gh_profiler/cli.py
@@ -1,12 +1,8 @@
 """Main CLI entry point for gh-profiler."""
 
-import json
-import sys
-
 import click
 
 from . import gh_profiler
-from .utils.infra_utils import run_cmd
 from .utils import cli_utils
 
 
@@ -15,7 +11,7 @@ from .utils import cli_utils
 @click.version_option(package_name="gh-profiler")
 def main(target):
     """Examine a GitHub user's profile, to help quickly decide how much to invest in their contributions.
-    
+
     You can target a GitHub username, or a PR/issue number from the repository you're working in.
 
     \b
@@ -36,7 +32,7 @@ def main(target):
         pr_issue_num = int(target)
     except ValueError:
         gh_profiler.main(target)
-    
+
     # The user provided a PR/issue number. Get the relevant username, then
     # call gh_profiler.main().
     username = cli_utils.get_username(pr_issue_num)