+++ src/gh_profiler/utils/profile_utils.py
@@ -1,16 +1,16 @@
 """Utils for retrieving user information."""
 
 import json
-import os
-import re
 from datetime import datetime as dt
 from datetime import timezone as tz
 from datetime import timedelta
 from urllib.parse import quote
+from textwrap import dedent
 import sys
 
 from .profile_data import profile_data as pdata
 from . import infra_utils
+
 
 def ensure_gh():
     """Make sure user has gh installed and is authenticated.
@@ -32,21 +32,23 @@ def ensure_gh():
         msg += "\n  If you're not authenticated, run `gh auth login`."
         sys.exit(msg)
 
+
 def get_profile_info():
     """Get all the profile info we'll need."""
     cmd = f"gh api users/{pdata.username} --jq '{{login, name, created_at, company, blog, location, email, bio}}'"
 
     profile_info = infra_utils.run_cmd(cmd)
-    
+
     pdata.profile_info = json.loads(profile_info)
 
     if "created_at" not in pdata.profile_info:
         sys.exit(f"GitHub user '{pdata.username}' not found.")
-    
+
     # On Linux, an invalid profile seems to return a dict with all the fields,
     # but every value is None.
     if pdata.profile_info["created_at"] is None:
         sys.exit(f"GitHub user '{pdata.username}' not found.")
+
 
 def get_pr_activity():
     """Get information about recent PR activity."""
@@ -69,3 +71,50 @@ def get_pr_activity():
         msg += "\n  You may want to try running the command again."
         sys.exit(msg)
 
+
+def get_issue_activity():
+    """Get target user's recent public issue activity."""
+    cutoff = (dt.now(tz.utc) - timedelta(days=21)).date().isoformat()
+    gh_call = _get_gh_issues_call(pdata.username, cutoff)
+    try:
+        issue_activity = infra_utils.run_cmd(gh_call)
+    except ValueError:
+        msg = "Couldn't get recent issue activity. The gh CLI may have timed out."
+        msg += "\n  You may want to try running the command again."
+        sys.exit(msg)
+
+    pdata.issue_activity = json.loads(issue_activity)["data"]["search"]
+
+
+# --- Helper functions ---
+
+
+def _get_gh_issues_call(username, cutoff):
+    """Return the gh call for recent public issue activity."""
+    gh_call = f"""
+        gh api graphql -f query='
+        query($q: String!, $n: Int!) {{
+        search(query: $q, type: ISSUE, first: $n) {{
+            issueCount
+            pageInfo {{
+            hasNextPage
+            endCursor
+            }}
+            nodes {{
+            ... on Issue {{
+                number
+                title
+                createdAt
+                state
+                stateReason
+                url
+                repository {{
+                nameWithOwner
+                }}
+            }}
+            }}
+        }}
+        }}' -F q='author:{username} is:issue is:public created:>={cutoff}' -F n=100
+    """
+
+    return dedent(gh_call).strip()