+++ src/gh_profiler/utils/summary_utils.py
@@ -9,6 +9,7 @@ def show_summary():
     summary = _get_summary()
     print(summary)
 
+
 def _get_summary():
     """Build a summary string.
 
@@ -24,10 +25,14 @@ def _get_summary():
     summary += "\n"
     summary += _pr_activity_summary()
     summary += "\n"
-    
+    summary += _issue_activity_summary()
+    summary += "\n"
+
     return summary.strip()
 
+
 # --- Helper functions ---
+
 
 def _profile_summary():
     """Summarize information from the user's profile dict."""
@@ -53,6 +58,7 @@ def _profile_summary():
 
     return summary
 
+
 def _bio_summary(bio):
     """Summarize bio section of profile."""
     if bio in (None, ""):
@@ -67,6 +73,7 @@ def _bio_summary(bio):
         summary += f"        {line}\n"
     return summary
 
+
 def _pr_activity_summary():
     """Summarize recent PR activity."""
     if pdata.opened_count < 10:
@@ -76,8 +83,28 @@ def _pr_activity_summary():
     # Only show merged if it's a good sign.
     if pdata.flag_merged_pr == flags.green_flag:
         summary += f"  {pdata.flag_merged_pr} {pdata.merged_count} of {pdata.opened_count} PRs have been merged in the last 21 days.\n"
-    
+
     # Include number closed for everyone.
     summary += f"  {pdata.flag_closed_pr} {pdata.closed_count} of {pdata.opened_count} PRs have been closed without merging in the last 21 days.\n"
+
+    return summary
+
+
+def _issue_activity_summary():
+    """Summarize recent public issue activity."""
+    ...
+    if pdata.new_issue_count == 0:
+        return f"\n    {pdata.usename} has not opened any new issues in the last 21 days.\n"
+
+    summary = f"  {pdata.flag_overall_issues} {pdata.username} has opened {pdata.new_issue_count} new issues in the last 21 days.\n"
+    summary += f"     {pdata.flag_issues_not_planned} {pdata.issues_not_planned} issues have been closed as NOT_PLANNED.\n"
+
+    # Repeated issues:
+    if pdata.total_repeats == 0:
+        summary += f"     {pdata.flag_repeated_issues} {pdata.total_repeats} issues were opened with the same title.\n"
+    else:
+        summary += f"     {pdata.flag_repeated_issues} {pdata.total_repeats} issues were opened with the same title:\n"
+    for title, count in pdata.repeated_issue_titles.items():
+        summary += f"        {title} ({count})\n"
 
     return summary