+++ tests/unit_tests/reference_files/reference_summaries.py
@@ -12,6 +12,10 @@ GitHub user: ehmatthes
      Empty fields: company, bio
 
   🟢 ehmatthes has opened fewer than 10 PRs in the last 21 days.
+
+  🟢 ehmatthes has opened 7 new issues in the last 21 days.
+     🟢 0 issues have been closed as NOT_PLANNED.
+     🟢 0 issues were opened with the same title.
 """.strip()
 
 summary_empty_profile = """
@@ -21,4 +25,8 @@ GitHub user: ehmatthes
   🔴 No profile information has been provided.
 
   🟢 ehmatthes has opened fewer than 10 PRs in the last 21 days.
+
+  🟢 ehmatthes has opened 7 new issues in the last 21 days.
+     🟢 0 issues have been closed as NOT_PLANNED.
+     🟢 0 issues were opened with the same title.
 """.strip()