+++ tests/unit_tests/test_summary.py
@@ -12,18 +12,20 @@ from reference_files import reference_summaries
 
 # --- Fixtures ---
 
+
 @pytest.fixture(autouse=True)
 def filled_pdata():
     """Fill in a basic pdata profile.
 
     This fixture doesn't need to be passed, because it modifies the singleton
     pdata object.
-    
+
     DEV: Consider processing the profile data, instead of setting it?
          Maybe that's an integration test?
          Maybe unit test the processing function?
     """
     pdata.username = "ehmatthes"
+
     pdata.profile_info = {
         "name": "Eric Matthes",
         "company": None,
@@ -32,16 +34,30 @@ def filled_pdata():
         "email": "ehmatthes@gmail.com",
         "bio": None,
     }
+    pdata.account_age = timedelta(days=5058)
+
     pdata.opened_count = 5
     pdata.closed_count = 1
     pdata.merged_count = 3
-    pdata.account_age = timedelta(days=5058)
+
     pdata.flag_age = flags.green_flag
     pdata.flag_profile = flags.green_flag
 
     # This is taken from analysis_utils.py:
     fields = ["name", "company", "blog", "location", "email", "bio"]
-    pdata.profile_dict = {field:pdata.profile_info[field] for field in fields}
+    pdata.profile_dict = {field: pdata.profile_info[field] for field in fields}
+
+    # Issue fields
+    pdata.new_issue_count = 7
+    pdata.total_repeats = 0
+    pdata.repeated_issue_titles = {}
+
+    pdata.flag_issues_not_planned = flags.green_flag
+    pdata.flag_repeated_issues = flags.green_flag
+
+    # Call function to process this?
+    pdata.flag_overall_issues = flags.green_flag
+
 
 @pytest.fixture()
 def empty_profile_info():
@@ -54,18 +70,19 @@ def empty_profile_info():
     pdata.flag_profile = flags.red_flag
 
 
-
 # --- Test functions ---
+
 
 def test_summary():
     """Test the overall summary output.
-    
+
     DEV: Consider processing the profile data, instead of setting it?
          Maybe that's an integration test?
          Maybe unit test the processing function?
     """
     summary = summary_utils._get_summary()
     assert summary.strip() == reference_summaries.full_summary
+
 
 def test_summary_empty_profile(empty_profile_info):
     """Test summary for user with an empty profile."""